+++ completion/git-machete.completion.bash
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 _git_machete() {
-    cmds="add anno d delete-unmanaged diff discover e edit file fork-point g go help l list log reapply show slide-out s status traverse update version"
+    cmds="add anno d delete-unmanaged diff discover e edit file fork-point g go help is-managed l list log reapply show slide-out s status traverse update version"
     help_topics="$cmds format hooks"
 
     categories="managed slidable slidable-after unmanaged with-overridden-fork-point"
@@ -11,7 +11,8 @@ _git_machete() {
     opt_start_from_args="here root first-root"
 
     common_opts="--debug -h --help -v --verbose --version"
-    add_opts="-o --onto= --yes"
+    add_opts="-o --onto= -R --as-root --yes"
+    anno_opts="-b --branch="
     delete_unmanaged_opts="--yes"
     diff_opts="-s --stat"
     discover_opts="-C --checked-out-since= -l --list-commits -r --roots= -y --yes"
@@ -23,16 +24,17 @@ _git_machete() {
     update_opts="-f --fork-point= -M --merge -n --no-edit-merge --no-interactive-rebase"
 
     case $cur in
+        --branch=*|--onto=*) __gitcomp_nl "$(git machete list managed)" "" "${cur##--*=}" ;;
         --checked-out-since=*) __gitcomp "" ;;
         --color=*) __gitcomp "$opt_color_args" "" "${cur##--color=}" ;;
         --down-fork-point=*|--fork-point=*|--override-to=*) __gitcomp "$(__git_refs)" "" "${cur##--*=}" ;;
-        --onto=*) __gitcomp_nl "$(git machete list managed)" "" "${cur##--onto=}" ;;
         --return-to=*) __gitcomp "$opt_return_to_args" "" "${cur##--return-to=}" ;;
         --roots=*) __gitcomp "$(__git_heads)" "" "${cur##--roots=}" ;;
         --start-from=*) __gitcomp "$opt_start_from_args" "" "${cur##--start-from=}" ;;
         -*)
             case ${COMP_WORDS[2]} in
                 add) __gitcomp "$common_opts $add_opts" ;;
+                anno) __gitcomp "$common_opts $anno_opts" ;;
                 d|diff) __gitcomp "$common_opts $diff_opts" ;;
                 delete-unmanaged) __gitcomp "$common_opts $delete_unmanaged_opts" ;;
                 discover) __gitcomp "$common_opts $discover_opts" ;;
@@ -50,13 +52,13 @@ _git_machete() {
              else
                 prev=${COMP_WORDS[COMP_CWORD-1]}
                 case $prev in
+                    -b|--branch|-o|--onto) __gitcomp_nl "$(git machete list managed)" ;;
                     -C|--checked-out-since) __gitcomp "" ;;
                     --color) __gitcomp "$opt_color_args" ;;
                     -d|--down-fork-point|-f|--fork-point|--override-to) __gitcomp "$(__git_refs)" ;;
                     # TODO (GH issue #25): We don't complete --help since it's going to be captured by git anyway
                     # (and results in redirection to yet non-existent man for `git-machete`).
                     -h) __gitcomp "$help_topics" ;;
-                    -o|--onto) __gitcomp_nl "$(git machete list managed)" ;;
                     --return-to) __gitcomp "$opt_return_to_args" ;;
                     # TODO complete the comma-separated list of roots
                     -r|--roots) __gitcomp "$(__git_heads)" ;;
@@ -65,8 +67,8 @@ _git_machete() {
                     *)
                         case ${COMP_WORDS[2]} in
                             add) __gitcomp_nl "$(git machete list unmanaged)" ;;
-                            d|diff|fork-point|l|log) __gitcomp "$(__git_heads)" ;;
-                            g|go|show) __gitcomp "$directions" ;;
+                            d|diff|fork-point|is-managed|l|log) __gitcomp "$(__git_heads)" ;;
+                            g|go) __gitcomp "$directions" ;;
                             help) __gitcomp "$help_topics" ;;
                             list)
                                 if [[ $COMP_CWORD -eq 3 ]]; then
@@ -74,13 +76,14 @@ _git_machete() {
                                 elif [[ $COMP_CWORD -eq 4 && $prev == slidable-after ]]; then
                                     __gitcomp_nl "$(git machete list slidable)"
                                 fi ;;
+                            show) __gitcomp "current $directions" ;;
                             slide-out)
                                 if [[ $COMP_CWORD -eq 3 ]]; then
                                     __gitcomp_nl "$(git machete list slidable)"
                                 else
                                     __gitcomp_nl "$(git machete list slidable-after "$prev" 2>/dev/null)"
                                 fi ;;
-                            *) __gitcomp "" ;;
+                            *) COMPREPLY=('') ;; # not perfect (kinda-completes an empty string), but at least local paths aren't completed by default
                         esac ;;
                 esac
             fi